Some restructuring and cleanup
